# api/openapi.yaml
openapi: 3.1.0
info:
  title: Organon Personal API
  version: 0.1.0
  summary: Headless Org-mode as your personal API.
  description: |
    Tasks and notes live in plain Org files; this API is the only way clients
    read and change them. Everything that has a meaning in Org (state changes,
    repeaters, deadline warnings, agenda membership) is computed by Org itself,
    and everything about notes (IDs, links, backlinks) by org-roam.

    **Dates and instants.** Planning dates (`scheduled`, `deadline`) are
    calendar dates with an optional wall-clock time in the instance's calendar
    zone (`GET /api/v1/meta` → `calendar_tz`); they carry no offset. Instants
    (`closed_at`) are RFC 3339 UTC.

    Behavior is specified in `openspec/` (capabilities task-lifecycle,
    task-recurrence, agenda-queries, time-model, api-access, knowledge-nodes,
    knowledge-index).
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:8080
security:
  - bearer: []
...
paths:
...
  /api/v1/meta:
    get:
      tags: [system]
      operationId: getMeta
      summary: Instance metadata
      responses:
        "200":
          description: Instance metadata.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Meta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/EngineUnavailable" }
  /api/v1/agenda:
    get:
      tags: [agenda]
      operationId: getAgenda
      summary: The Org agenda for one day
      description: |
        Every entry Org Agenda shows for the day when that day is the current
        day: deadline warnings and overdue items apply to it.
      parameters:
        - $ref: "#/components/parameters/date"
      responses:
        "200":
          description: Agenda entries in agenda order.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AgendaList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Invalid" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/EngineUnavailable" }
  /api/v1/tasks/today:
    get:
      tags: [tasks]
      operationId: listTodayTasks
      summary: Open tasks on the day's agenda
      parameters:
        - $ref: "#/components/parameters/date"
      responses:
        "200": { $ref: "#/components/responses/TaskListWithAgenda" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Invalid" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/EngineUnavailable" }
  /api/v1/tasks/overdue:
    get:
      tags: [tasks]
      operationId: listOverdueTasks
      summary: Tasks on the day's agenda that are past-scheduled or past their deadline
      parameters:
        - $ref: "#/components/parameters/date"
      responses:
        "200": { $ref: "#/components/responses/TaskListWithAgenda" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Invalid" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/EngineUnavailable" }
  /api/v1/tasks/waiting:
    get:
      tags: [tasks]
      operationId: listWaitingTasks
      summary: All WAITING tasks, with or without dates
      responses:
        "200": { $ref: "#/components/responses/TaskList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/EngineUnavailable" }
  /api/v1/tasks/completed:
    get:
      tags: [tasks]
      operationId: listCompletedTasks
      summary: Tasks completed on a day, including repeating tasks
      parameters:
        - $ref: "#/components/parameters/date"
      responses:
        "200": { $ref: "#/components/responses/TaskList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Invalid" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/EngineUnavailable" }
  /api/v1/tasks:
    get:
      tags: [tasks]
      operationId: listTasks
      summary: Tasks by state, project and tag, with or without dates
      description: |
        Tasks in `org/tasks/` and `org/projects/`, in file order. Without
        `state`, the open states (TODO, NEXT, DOING, WAITING) are listed.
        Filters combine with AND; `tag` matches the task's own tags.
      parameters:
        - name: state
          in: query
          required: false
          description: Comma-separated states, e.g. `NEXT` or `DONE,CANCELLED`.
          schema:
            type: string
            pattern: "^(TODO|NEXT|DOING|WAITING|DONE|CANCELLED)(,(TODO|NEXT|DOING|WAITING|DONE|CANCELLED))*$"
        - name: project
          in: query
          required: false
          schema: { $ref: "#/components/schemas/ID" }
        - name: tag
          in: query
          required: false
          schema: { type: string, pattern: "^[A-Za-z0-9_@#%]+$" }
      responses:
        "200": { $ref: "#/components/responses/TaskList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Invalid" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/EngineUnavailable" }
    post:
      tags: [tasks]
      operationId: createTask
      summary: Create a task in the inbox or under a project
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTask" }
      responses:
        "201":
          description: The created task.
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Invalid" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/EngineUnavailable" }
  /api/v1/tasks/{id}:
    get:
      tags: [tasks]
      operationId: getTask
      summary: A task by its stable ID
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: The task.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Invalid" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/EngineUnavailable" }

    patch:
      tags: [tasks]
      operationId: updateTask
      summary: Change some fields of a task
      description: |
        Only the fields present change. `null` clears `priority`,
        `scheduled`, `deadline` and `repeat_to_state`. A planning date that is
        sent replaces the old one entirely, repeater and warning included.
        `expected_version` (the task's `version`) is required.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateTask" }
      responses:
        "200":
          description: The task after the change.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Invalid" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/EngineUnavailable" }
  /api/v1/tasks/{id}/{action}:
    post:
      tags: [tasks]
      operationId: transitionTask
      summary: Change a task's state
      description: |
        - `start` → DOING, `wait` → WAITING, `complete` → DONE.
        - `skip` → CANCELLED for this occurrence; a repeating task moves to
          its next occurrence.
        - `cancel` → CANCELLED and, for a repeating task, ends the series.
        - `todo` → TODO, `next` → NEXT; on a closed task this reopens it.

        `expected_state` is always required. `expected_version` (the task's
        `version`) is required for repeating tasks, which return to an open
        state after completion: without it a retried request could not be
        told apart from a new one.
      parameters:
        - $ref: "#/components/parameters/id"
        - name: action
          in: path
          required: true
          schema:
            type: string
            enum: [start, wait, complete, skip, cancel, todo, next]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Transition" }
      responses:
        "200":
          description: The task after the transition.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TransitionResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Invalid" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "503": { $ref: "#/components/responses/EngineUnavailable" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: |
        Tokens are created with `organon token new` and carry scopes:
        `read` for every GET, `tasks:write` for creating and changing tasks
        and projects, `nodes:write` for creating nodes. No scope implies
        another.
...
  responses:
    TaskList:
      description: Tasks.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/TaskList" }
...
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: The token lacks the required scope.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    RateLimited:
      description: Too many failed authentication attempts from this address.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Meta:
      type: object
      required: [calendar_tz, today, doing_limit, version, engine_version]
      properties:
        calendar_tz: { type: string, examples: ["Asia/Seoul"] }
        today: { $ref: "#/components/schemas/Date" }
        doing_limit: { type: integer, minimum: 1 }
        version: { type: string }
        engine_version: { type: string }
...